# docs/swagger.yaml
definitions:
...
  github_com_michaelwp_student_attendance_internal_models.Class:
    properties:
      created_at:
        type: string
      description:
        type: string
      homeroom_teacher:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /classes:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all classes
      parameters:
      - default: 10
        description: Number of classes to return (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of classes to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of classes retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all classes
      tags:
      - Classes
    post:
      consumes:
      - application/json
      description: Create a new class in the system
      parameters:
      - description: Class data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.Class'
      produces:
      - application/json
      responses:
        "201":
          description: Class created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new class
      tags:
      - Classes
  /classes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a class from the system
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Class deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid class ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete class
      tags:
      - Classes
    get:
      consumes:
      - application/json
      description: Retrieve a specific class by ID
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Class retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid class ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Class not found
          schema:
            additionalProperties: true
            type: object
      summary: Get class by ID
      tags:
      - Classes
    put:
      consumes:
      - application/json
      description: Update a class's information
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      - description: Class data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/github_com_michaelwp_student_attendance_internal_models.Class'
      produces:
      - application/json
      responses:
        "200":
          description: Class updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update class
      tags:
      - Classes
  /classes/teacher-id/{teacherId}:
    get:
      consumes:
      - application/json
      description: Retrieve all classes assigned to a specific teacher
      parameters:
      - description: Teacher ID
        in: path
        name: teacherId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Classes retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Teacher ID is required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get classes by teacher
      tags:
      - Classes
...
swagger: "2.0"